fix: project structure baseline
