+++ src/sections/About.tsx
@@ -1,0 +1,3 @@
+export const AboutSection = () => {
+  return <div>About Section</div>;
+};
