+++ src/sections/Contact.tsx
@@ -1,0 +1,3 @@
+export const ContactSection = () => {
+  return <div>Contact Section</div>;
+};
