+++ src/sections/Hero.tsx
@@ -1,0 +1,3 @@
+export const HeroSection = () => {
+  return <div>Hero Section</div>;
+};
