+++ src/sections/Tape.tsx
@@ -1,0 +1,3 @@
+export const TapeSection = () => {
+  return <div>Tape Section</div>;
+};
